React Router Dom

+++ backend/src/controllers/notes.ts
@@ -2,10 +2,13 @@ import { RequestHandler } from "express";
 import NoteModel from "../models/note";
 import createHttpError from "http-errors";
 import mongoose from "mongoose";
+import { assertIsDefine } from "../util/assertIsDefine";
 
 export const getNotes: RequestHandler = async (req, res, next) => {
+  const authUserId = req.session.userId;
   try {
-    const notes = await NoteModel.find().exec();
+    assertIsDefine(authUserId);
+    const notes = await NoteModel.find({ userId: authUserId }).exec();
     res.status(200).json(notes);
   } catch (error) {
     next(error);
@@ -14,14 +17,19 @@ export const getNotes: RequestHandler = async (req, res, next) => {
 
 export const getNote: RequestHandler = async (req, res, next) => {
   const id = req.params.id;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefine(authUserId);
     if (!mongoose.isValidObjectId(id)) {
       throw createHttpError(400, "Invalid ID!");
     }
     const note = await NoteModel.findById(id).exec();
     if (!note) {
       throw createHttpError(404, "Note not found!");
+    }
+    if (!note.userId.equals(authUserId)) {
+      throw createHttpError(401, "Unauthorized | Not your note!");
     }
     res.status(200).json(note);
   } catch (error) {
@@ -42,12 +50,14 @@ export const createNote: RequestHandler<
 > = async (req, res, next) => {
   const title = req.body.title;
   const text = req.body.text;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefine(authUserId);
     if (!title) {
       throw createHttpError(400, "Title is required!");
     }
-    const newNote = new NoteModel({ title, text });
+    const newNote = new NoteModel({ userId: authUserId, title, text });
     const note = await newNote.save();
     res.status(201).json(note);
   } catch (error) {
@@ -73,8 +83,11 @@ export const updateNote: RequestHandler<
   const id = req.params.id;
   const newTitle = req.body.title;
   const newText = req.body.text;
+  const authUserId = req.session.userId;
 
   try {
+    assertIsDefine(authUserId);
+
     if (!mongoose.isValidObjectId(id)) {
       throw createHttpError(400, "Invalid ID!");
     }
@@ -87,6 +100,10 @@ export const updateNote: RequestHandler<
 
     if (!note) {
       throw createHttpError(404, "Note not found!");
+    }
+
+    if (!note.userId.equals(authUserId)) {
+      throw createHttpError(401, "Unauthorized | Not your note!");
     }
 
     note.title = newTitle;
@@ -102,7 +119,10 @@ export const updateNote: RequestHandler<
 
 export const deleteNote: RequestHandler = async (req, res, next) => {
   const id = req.params.id;
+  const authUserId = req.session.userId;
+
   try {
+    assertIsDefine(authUserId);
     if (!mongoose.isValidObjectId(id)) {
       throw createHttpError(400, "Invalid ID!");
     }
@@ -111,6 +131,10 @@ export const deleteNote: RequestHandler = async (req, res, next) => {
 
     if (!note) {
       throw createHttpError(404, "Note not found!");
+    }
+
+    if (!note.userId.equals(authUserId)) {
+      throw createHttpError(401, "Unauthorized | Not your note!");
     }
 
     await note.deleteOne();
